tcb property and 64 bit values

+++ include/support.h
@@ -116,7 +116,7 @@ bool get_random(int num_bits, byte* out);
 
 bool make_key_entity(const key_message& key, entity_message* ent);
 bool make_measurement_entity(const string& measurement, entity_message* ent);
-bool make_property(string& name, string& type, string& cmp, int int_value,
+bool make_property(string& name, string& type, string& cmp, uint64_t int_value,
       string& string_value, property* prop);
 bool make_platform(const string& type, const properties& p, const key_message* at,
       platform* plat);
